requirements (1).txt
streamlit
pillow
pandas
requests
gspread
pytz
streamlit-lottie
streamlit-extras
oauth2client
google-auth
